fix(tests): split harness suites under file cap
