+++ tests/unit/harness/source-candidate.test.ts
@@ -417,244 +417,4 @@ describe('harness source candidate validation', () => {
 		expect(result.candidate.files[0]?.beforeSha256).not.toBe(sha256(dirtyText));
 		expect(result.candidate.files[0]?.afterSha256).toBe(sha256(patchedText));
 	});
-
-	it('binds declared Git object ids and preserves exact EOF newline semantics', async () => {
-		const forgedOld = await validate({
-			patch: [
-				'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-				`index ${'f'.repeat(7)}..${gitBlobId('export const value = 2;\nexport const name = "demo";\n').slice(0, 7)} 100644`,
-				'--- a/src/agents/demo.ts',
-				'+++ b/src/agents/demo.ts',
-				'@@ -1,2 +1,2 @@',
-				'-export const value = 1;',
-				'+export const value = 2;',
-				' export const name = "demo";',
-				'',
-			].join('\n'),
-		});
-		expect(forgedOld.ok).toBe(false);
-		if (forgedOld.ok) throw new Error('expected rejection');
-		expect(forgedOld.code).toBe('STALE_BASE');
-
-		const forgedNew = await validate({
-			patch: [
-				'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-				`index ${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}..${'f'.repeat(7)} 100644`,
-				'--- a/src/agents/demo.ts',
-				'+++ b/src/agents/demo.ts',
-				'@@ -1,2 +1,2 @@',
-				'-export const value = 1;',
-				'+export const value = 2;',
-				' export const name = "demo";',
-				'',
-			].join('\n'),
-		});
-		expect(forgedNew.ok).toBe(false);
-		if (forgedNew.ok) throw new Error('expected rejection');
-		expect(forgedNew.code).toBe('PATCH_DOES_NOT_APPLY');
-
-		const noNewlineText = 'export const created = true;';
-		const noNewline = await validate({
-			patch: [
-				'diff --git a/src/agents/no-newline.ts b/src/agents/no-newline.ts',
-				'new file mode 100644',
-				`index ${'0'.repeat(7)}..${gitBlobId(noNewlineText).slice(0, 7)}`,
-				'--- /dev/null',
-				'+++ b/src/agents/no-newline.ts',
-				'@@ -0,0 +1 @@',
-				'+export const created = true;',
-				'\\ No newline at end of file',
-				'',
-			].join('\n'),
-		});
-		expect(noNewline.ok).toBe(true);
-		if (!noNewline.ok) throw new Error(noNewline.reason);
-		expect(noNewline.candidate.files[0]?.afterSha256).toBe(
-			sha256(noNewlineText),
-		);
-
-		const earlierBase = 'export const first = 1;\nexport const last = true;';
-		const earlierAfter = 'export const first = 2;\nexport const last = true;';
-		gitEntries.set('src/agents/no-final-newline.ts', {
-			mode: '100644',
-			objectId: gitBlobId(earlierBase),
-		});
-		gitBlobs.set('src/agents/no-final-newline.ts', Buffer.from(earlierBase));
-		const earlierHunk = await validate({
-			patch: [
-				'diff --git a/src/agents/no-final-newline.ts b/src/agents/no-final-newline.ts',
-				`index ${gitBlobId(earlierBase).slice(0, 7)}..${gitBlobId(earlierAfter).slice(0, 7)} 100644`,
-				'--- a/src/agents/no-final-newline.ts',
-				'+++ b/src/agents/no-final-newline.ts',
-				'@@ -1 +1 @@',
-				'-export const first = 1;',
-				'+export const first = 2;',
-				'',
-			].join('\n'),
-		});
-		expect(earlierHunk.ok).toBe(true);
-		if (!earlierHunk.ok) throw new Error(earlierHunk.reason);
-		expect(earlierHunk.candidate.files[0]?.afterSha256).toBe(
-			sha256(earlierAfter),
-		);
-	});
-
-	it('rejects malformed, miscounted, and overlapping hunks', async () => {
-		for (const [patch, code] of [
-			[
-				[
-					'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-					`index ${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}..${gitBlobId('export const value = 2;\nexport const name = "demo";\n').slice(0, 7)} 100644`,
-					'--- a/src/agents/demo.ts',
-					'+++ b/src/agents/demo.ts',
-					'@@ -1,1 +1,2 @@',
-					'-export const value = 1;',
-					'+export const value = 2;',
-					'',
-				].join('\n'),
-				'UNSUPPORTED_PATCH',
-			],
-			[
-				[
-					'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-					`index ${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}..${gitBlobId('export const value = 2;\nexport const name = "demo";\n').slice(0, 7)} 100644`,
-					'--- a/src/agents/demo.ts',
-					'+++ b/src/agents/demo.ts',
-					'@@ -1,1 +1,1 @@',
-					'-export const value = 1;',
-					'+export const value = 2;',
-					'@@ -1,1 +1,1 @@',
-					'-export const name = "demo";',
-					'+export const name = "dirty";',
-					'',
-				].join('\n'),
-				'UNSUPPORTED_PATCH',
-			],
-			[
-				[
-					'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-					`index ${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}..${gitBlobId('export const value = 2;\nexport const name = "demo";\n').slice(0, 7)} 100644`,
-					'--- a/src/agents/demo.ts',
-					'+++ b/src/agents/demo.ts',
-					'@@ -1,1 +2,1 @@',
-					'-export const value = 1;',
-					'+export const value = 2;',
-					'',
-				].join('\n'),
-				'PATCH_DOES_NOT_APPLY',
-			],
-		] as const) {
-			const result = await validate({ patch });
-			expect(result.ok).toBe(false);
-			if (result.ok) throw new Error('expected rejection');
-			expect(result.code).toBe(code);
-		}
-	});
-
-	it('accepts git-quoted paths safely', async () => {
-		const quotedPath = 'src/agents/space path.ts';
-		const secondQuotedPath = 'src/agents/second path.ts';
-		mkdirSync(path.join(root, 'src', 'agents'), { recursive: true });
-		writeFileSync(
-			path.join(root, 'src', 'agents', 'space path.ts'),
-			'export const label = "space";\n',
-			'utf8',
-		);
-		gitEntries.set(quotedPath, {
-			mode: '100644',
-			objectId: gitBlobId('export const label = "space";\n'),
-		});
-		gitBlobs.set(quotedPath, Buffer.from('export const label = "space";\n'));
-		gitEntries.set(secondQuotedPath, {
-			mode: '100644',
-			objectId: gitBlobId('export const second = 1;\n'),
-		});
-		gitBlobs.set(secondQuotedPath, Buffer.from('export const second = 1;\n'));
-
-		const result = await validate({
-			patch: [
-				'diff --git "a/src/agents/space path.ts" "b/src/agents/space path.ts"',
-				`index ${gitBlobId('export const label = "space";\n').slice(0, 7)}..${gitBlobId('export const label = "patched";\n').slice(0, 7)} 100644`,
-				'--- "a/src/agents/space path.ts"',
-				'+++ "b/src/agents/space path.ts"',
-				'@@ -1 +1 @@',
-				'-export const label = "space";',
-				'+export const label = "patched";',
-				'diff --git "a/src/agents/second path.ts" "b/src/agents/second path.ts"',
-				`index ${gitBlobId('export const second = 1;\n').slice(0, 7)}..${gitBlobId('export const second = 2;\n').slice(0, 7)} 100644`,
-				'--- "a/src/agents/second path.ts"',
-				'+++ "b/src/agents/second path.ts"',
-				'@@ -1 +1 @@',
-				'-export const second = 1;',
-				'+export const second = 2;',
-				'',
-			].join('\n'),
-		});
-		expect(result.ok).toBe(true);
-		if (!result.ok) throw new Error(result.reason);
-		expect(result.candidate.files[0]?.relativePath).toBe(quotedPath);
-		expect(result.candidate.files[1]?.relativePath).toBe(secondQuotedPath);
-	});
-
-	it('admits inert add, delete, rename, and mode-only patch forms without mutation', async () => {
-		const admittedPatches = [
-			[
-				'diff --git a/src/agents/new.ts b/src/agents/new.ts',
-				'new file mode 100644',
-				`index ${'0'.repeat(7)}..${gitBlobId('export const created = true;\n').slice(0, 7)}`,
-				'--- /dev/null',
-				'+++ b/src/agents/new.ts',
-				'@@ -0,0 +1 @@',
-				'+export const created = true;',
-				'',
-			].join('\n'),
-			[
-				'diff --git a/src/agents/extra.ts b/src/agents/extra.ts',
-				'deleted file mode 100644',
-				`index ${gitBlobId('export const extra = true;\n').slice(0, 7)}..${'0'.repeat(7)}`,
-				'--- a/src/agents/extra.ts',
-				'+++ /dev/null',
-				'@@ -1 +0,0 @@',
-				'-export const extra = true;',
-				'',
-			].join('\n'),
-			[
-				'diff --git a/src/agents/demo.ts b/src/agents/demo-renamed.ts',
-				'similarity index 100%',
-				'rename from src/agents/demo.ts',
-				'rename to src/agents/demo-renamed.ts',
-				'',
-			].join('\n'),
-			[
-				'diff --git a/src/agents/demo.ts b/src/agents/demo-copy.ts',
-				'similarity index 100%',
-				'copy from src/agents/demo.ts',
-				'copy to src/agents/demo-copy.ts',
-				'',
-			].join('\n'),
-			[
-				'diff --git a/src/agents/demo.ts b/src/agents/demo.ts',
-				'old mode 100644',
-				'new mode 100755',
-				`index ${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}..${gitBlobId('export const value = 1;\nexport const name = "demo";\n').slice(0, 7)}`,
-				'--- a/src/agents/demo.ts',
-				'+++ b/src/agents/demo.ts',
-				'',
-			].join('\n'),
-		];
-
-		const expectedOperations = ['add', 'delete', 'rename', 'copy', 'mode'];
-		for (const [index, admittedPatch] of admittedPatches.entries()) {
-			const result = await validate({ patch: admittedPatch });
-			expect(result.ok).toBe(true);
-			if (!result.ok) throw new Error(`${result.code}: ${result.reason}`);
-			expect(result.candidate.files[0]?.operation).toBe(
-				expectedOperations[index],
-			);
-		}
-
-		expect(
-			readFileSync(path.join(root, 'src', 'agents', 'demo.ts'), 'utf8'),
-		).toContain('value = 1');
-	});
 });
