+++ tests/unit/harness/store.test.ts
@@ -467,81 +467,4 @@ describe('harness durable store', () => {
 			await loadHarnessVersion(root, secondActivation.version.versionId),
 		).not.toBeNull();
 	});
-
-	it('serves recent history and allows new mutations after the ledger outgrows the replay window', async () => {
-		for (const id of ['candidate-window-a', 'candidate-window-b']) {
-			const stored = candidate(id, id[0]!.repeat(64));
-			await recordHarnessCandidate({ directory: root, candidate: stored });
-			await issueWriteApprovalFact({
-				directory: root,
-				request: buildHarnessActivationApprovalRequest({
-					targetSessionId: 'session-a',
-					candidate: stored,
-					...activationBinding(await loadHarnessCurrent(root), stored),
-				}),
-				issuingSessionId: 'human-session',
-			});
-			const activated = await activateHarnessCandidate({
-				directory: root,
-				candidateId: stored.candidate.candidateId,
-				consumerSessionId: 'session-a',
-				...activationBinding(await loadHarnessCurrent(root), stored),
-				maxReplayRecords: 2,
-			});
-			if (activated.status !== 'activated') {
-				throw new Error('expected activation');
-			}
-		}
-
-		const third = candidate('candidate-window-c', 'c'.repeat(64));
-		await recordHarnessCandidate({ directory: root, candidate: third });
-		await issueWriteApprovalFact({
-			directory: root,
-			request: buildHarnessActivationApprovalRequest({
-				targetSessionId: 'session-a',
-				candidate: third,
-				...activationBinding(await loadHarnessCurrent(root), third),
-			}),
-			issuingSessionId: 'human-session',
-		});
-		const activated = await activateHarnessCandidate({
-			directory: root,
-			candidateId: third.candidate.candidateId,
-			consumerSessionId: 'session-a',
-			...activationBinding(await loadHarnessCurrent(root), third),
-			maxReplayRecords: 2,
-		});
-		expect(activated.status).toBe('activated');
-
-		const page = await listHarnessHistory(root, {
-			limit: 1,
-			maxReplayRecords: 1,
-		});
-		expect(page.totalRecordCount).toBe(1);
-		expect(page.records).toHaveLength(1);
-		expect(page.records[0]?.kind).toBe('compacted');
-	});
-
-	it('returns newest-first bounded history pages and full chronological audits', async () => {
-		for (const id of ['candidate-audit-a', 'candidate-audit-b']) {
-			await recordHarnessCandidate({
-				directory: root,
-				candidate: candidate(id),
-			});
-		}
-		const page = await listHarnessHistory(root, { limit: 1 });
-		expect(page.totalRecordCount).toBe(2);
-		expect(page.records).toHaveLength(1);
-		expect(page.records[0]?.candidateId).toBe('candidate-audit-b');
-		const audit = await auditHarnessLedger(root, {
-			maxReplayRecords: 10,
-			maxSegments: 10,
-		});
-		expect(audit.outcome).toBe('ok');
-		if (audit.outcome !== 'ok') throw new Error('expected audit');
-		expect(audit.records.map((record) => record.candidateId)).toEqual([
-			'candidate-audit-a',
-			'candidate-audit-b',
-		]);
-	});
 });
